chore: Rename Title Component
