+++ src/component/@share/atom/index.ts
@@ -1,4 +1,4 @@
-export * from './temp';
+export * from './Title';
 export * from './buttonLarge';
 export * from './buttonSmall';
 export * from './ActiveBar';
